Extracted some code from structs.h
