+++ src/engine/structs/structs.h
@@ -188,72 +188,6 @@ struct TextBlocksQueue {
 	struct TextBlock *head = nullptr;
 	struct TextBlock *tail = nullptr;
 };
-// ===============================================================
-
-#ifndef HAVE_ZLIB
-struct z_stream;
-#endif
-
-struct IndexData {
-	IndexData() : vnum(0), total_online(0), stored(0), func(nullptr), farg(nullptr), proto(nullptr), zone(0), set_idx(-1) {}
-	explicit IndexData(int _vnum)
-		: vnum(_vnum), total_online(0), stored(0), func(nullptr), farg(nullptr), proto(nullptr), zone(0), set_idx(-1) {}
-
-	Vnum vnum;            // virtual number of this mob/obj       //
-	int total_online;        // number of existing units of this mob/obj //
-	int stored;        // number of things in rent file            //
-	int (*func)(CharData *, void *, int, char *);
-	char *farg;        // string argument for special function     //
-	Trigger *proto;    // for triggers... the trigger     //
-	int zone;            // mob/obj zone rnum //
-	size_t set_idx; // индекс сета в obj_sets::set_list, если != -1
-};
-
-// ===============================================================
-
-enum ECase {
-	kNom = 0,
-	kGen,
-	kDat,
-	kAcc,
-	kIns,
-	kPre,
-	kFirstCase = kNom,
-	kLastCase = kPre,
-};
-
-namespace parser_wrapper {
-// forward declaration
-class DataNode;
-}
-
-namespace base_structs {
-
-class ItemName {
- public:
-	ItemName();
-
-	using Ptr = std::unique_ptr<ItemName>;
-	using NameCases = std::array<std::string, ECase::kLastCase + 1>;
-	ItemName(ItemName &&i) noexcept;
-	ItemName &operator=(ItemName &&i) noexcept;
-
-	[[nodiscard]] const std::string &GetSingular(ECase name_case = ECase::kNom) const;
-	[[nodiscard]] const std::string &GetPlural(ECase name_case = ECase::kNom) const;
-	static Ptr Build(parser_wrapper::DataNode &node);
-
- private:
-	NameCases singular_names_;
-	NameCases plural_names_;
-};
-
-}
-//спецпроцедуры
-int exchange(CharData *ch, void *me, int cmd, char *argument);
-int horse_keeper(CharData *ch, void *me, int cmd, char *argument);
-int torc(CharData *ch, void *me, int cmd, char *argument);
-int mercenary(CharData *ch, void * /*me*/, int cmd, char *argument);
-int shop_ext(CharData *ch, void *me, int cmd, char *argument);
 
 #endif // STRUCTS_STRUCTS_H_
