+++ src/engine/ui/cmd_god/stat.cpp
@@ -24,6 +24,7 @@
 #include "gameplay/mechanics/stable_objs.h"
 #include "gameplay/economics/ext_money.h"
 #include "administration/proxy.h"
+#include "gameplay/ai/spec_procs.h"
 
 extern char *diag_weapon_to_char(const CObjectPrototype *obj, int show_wear);
